+++ docs/make_apidoc.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-sphinx-apidoc --no-toc -e -f -o . ../pyroomacoustics ../pyroomacoustics/version.py ../pyroomacoustics/c_package
+sphinx-apidoc --no-toc -e -o . ../pyroomacoustics ../pyroomacoustics/version.py ../pyroomacoustics/c_package