+++ examples/bss_live.py
@@ -2,22 +2,37 @@
 Live Demonstration of Blind Source Separation
 =============================================
 
-Demonstrate how to do blind source separation (BSS) using the indpendent vector
-analysis technique. The method implemented is described in the following
-publication.
+Demonstrate the performance of different blind source separation (BSS) algorithms:
+
+1) Independent Vector Analysis (IVA)
+The method implemented is described in the following publication.
 
     N. Ono, *Stable and fast update rules for independent vector analysis based
     on auxiliary function technique*, Proc. IEEE, WASPAA, 2011.
 
-It works in the STFT domain. The test files were extracted from the
-`CMU ARCTIC <http://www.festvox.org/cmu_arctic/>`_ corpus.
+2) Independent Low-Rank Matrix Analysis (ILRMA)
+The method implemented is described in the following publications
 
-Running this script will do two things.
+    D. Kitamura, N. Ono, H. Sawada, H. Kameoka, H. Saruwatari, *Determined blind
+    source separation unifying independent vector analysis and nonnegative matrix
+    factorization,* IEEE/ACM Trans. ASLP, vol. 24, no. 9, pp. 1626-1641, September 2016
 
-1. It will separate the sources.
-2. Show a plot of the clean and separated spectrograms
-3. Show a plot of the SDR and SIR as a function of the number of iterations.
-4. Create a `play(ch)` function that can be used to play the `ch` source (if you are in ipython say).
+    D. Kitamura, N. Ono, H. Sawada, H. Kameoka, and H. Saruwatari *Determined Blind
+    Source Separation with Independent Low-Rank Matrix Analysis*, in Audio Source Separation,
+    S. Makino, Ed. Springer, 2018, pp.  125-156.
+
+Both algorithms work in the STFT domain. The test files are recorded by an external
+microphone array.
+
+Depending on the input arguments running this script will do these actions:.
+
+1. Record source signals from a connected microphone array
+2. Separate the sources.
+3. Show a plot of the clean and separated spectrograms
+4. Show a plot of the SDR and SIR as a function of the number of iterations.
+5. Create a `play(ch)` function that can be used to play the `ch` source (if you are in ipython say).
+6. Save the separated sources as .wav files
+7. Show a GUI where a mixed signals and the separated sources can be played
 
 This script requires the `sounddevice` packages to run.
 '''
@@ -36,10 +51,14 @@ import sounddevice as sd
 
 if __name__ == '__main__':
 
+    choices = ['ilrma', 'auxiva']
+
     import argparse
     parser = argparse.ArgumentParser(description='Records a segment of speech and then performs separation')
     parser.add_argument('-b', '--block', type=int, default=2048,
-            help='STFT block length')
+            help='STFT block size')
+    parser.add_argument('-a', '--algo', type=str, default=choices[0], choices=choices,
+            help='Chooses BSS method to run')
     parser.add_argument('-D', '--device', type=int,
             help='The sounddevice recording device id (obtain it with `python -m sounddevice`)')
     parser.add_argument('-d', '--duration', type=float,
@@ -80,8 +99,15 @@ if __name__ == '__main__':
         print('  AuxIVA Iter', it)
         it += 10
 
-    # Run AuxIVA
-    Y = pra.bss.auxiva(X, n_iter=args.n_iter, proj_back=True, callback=cb_print)
+    # Run BSS
+    bss_type = args.algo
+    if bss_type == 'auxiva':
+        # Run AuxIVA
+        Y = pra.bss.auxiva(X, n_iter=args.n_iter, proj_back=True, callback=cb_print)
+    elif bss_type == 'ilrma':
+        # Run ILRMA
+        Y = pra.bss.ilrma(X, n_iter=args.n_iter, n_components=30, proj_back=True,
+            callback=cb_print)
 
     # run iSTFT
     y = np.array([pra.istft(Y[:,:,ch], L, L, transform=np.fft.irfft, zp_front=L//2, zp_back=L//2) for ch in range(Y.shape[2])])