+++ pyroomacoustics/room.py
@@ -1,6 +1,174 @@
 # @version: 1.0  date: 05/06/2015 by Sidney Barthe
 # @author: robin.scheibler@epfl.ch, ivan.dokmanic@epfl.ch, sidney.barthe@epfl.ch
 # @copyright: EPFL-IC-LCAV 2015
+'''
+Room
+====
+
+The three main classes are :py:obj:`pyroomacoustics.room.Room`,
+:py:obj:`pyroomacoustics.soundsource.SoundSource`, and
+:py:obj:`pyroomacoustics.beamforming.MicrophoneArray`. On a high level, a
+simulation scenario is created by first defining a room to which a few sound
+sources and a microphone array are attached. The actual audio is attached to
+the source as raw audio samples. The image source method (ISM) is then used to
+find all image sources up to a maximum specified order and room impulse
+responses (RIR) are generated from their positions. The microphone signals are
+then created by convolving the audio samples associated to sources with the
+appropriate RIR. Since the simulation is done on discrete-time signals, a
+sampling frequency is specified for the room and the sources it contains.
+Microphones can optionally operate at a different sampling frequency; a rate
+conversion is done in this case.
+
+Simulating a Shoebox Room
+-------------------------
+
+We will first walk through the steps to simulate a shoebox-shaped room in 3D.
+
+
+Create the room
+~~~~~~~~~~~~~~~
+
+So-called shoebox rooms are pallelepipedic rooms with 4 or 6 walls (in 2D and 3D,
+respectiely), all at right angles. They are defined by a single vector that contains
+the lengths of the walls. They have the advantage of being simple to define and very
+efficient to simulate. A ``9m x 7.5m x 3.5m`` room is simply defined like this
+
+.. code-block:: python
+
+    import pyroomacoustics as pra
+    room = pra.ShoeBox([9, 7.5, 3.5], fs=16000, absorption=0.35, max_order=17)
+
+The second argument is the sampling frequency at which the RIR will be
+generated. Note that the default value of ``fs`` is 8 kHz. The third argument
+is the absorption of the walls, namely reflections are multiplied by ``(1 -
+absorption)`` for every wall they hit. The fourth argument is the maximum
+number of reflections allowed in the ISM.
+
+The relationship between ``absorption``/``max_order`` and `reverberation time
+<https://en.wikipedia.org/wiki/Reverberation>`_ (the T60 or RT60 in the
+acoustics literature) is not straightforward. `Sabine's formula
+<https://en.wikipedia.org/wiki/Reverberation#Sabine_equation>`_ can be used to
+some extent to set these parameters.
+
+
+Add sources and microphones
+~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Sources are fairly straighforward to create. They take their location as single
+mandatory argument, and a signal and start time as optional arguments.  Here we
+create a source located at ``[2.5, 3.73, 1.76]`` within the room, that will utter
+the content of the wav file ``speech.wav`` starting at ``1.3 s`` into the simulation.
+
+.. code-block:: python
+
+    # import a mono wavfile as the source signal
+    # the sampling frequency should match that of the room
+    from scipy.io import wavfile
+    _, audio = wavfile.read('speech.wav')
+
+    my_source = pra.SoundSource([2.5, 3.73, 1.76], signal=audio, delay=1.3)
+
+    # place the source in the room
+    room.add_source(my_source)
+
+The locations of the microphones in the array should be provided in a numpy
+``nd-array`` of size ``(ndim, nmics)``, that is each column contains the
+coordinates of one microphone. This array is used to construct a
+:py:obj:`pyroomacoustics.beamforming.MicrophoneArray` object, together with the
+sampling frequency for the microphone. Note that it can be different from that
+of the room, in which case resampling will occur. Here, we create an array
+with two microphones placed at ``[6.3, 4.87, 1.2]`` and ``[6.3, 4.93, 1.2]``.
+
+.. code-block:: python
+
+    # define the location of the array
+    import numpy as np
+    R = np.c_[
+        [6.3, 4.87, 1.2],  # mic 1
+        [6.3, 4.93, 1.2],  # mic 2
+        ]
+
+    # the fs of the microphones is the same as the room
+    mic_array = pra.MicrophoneArray(R, room.fs)
+
+    # finally place the array in the room
+    room.add_microphone_array(mic_array)
+
+A number of routines exist to create regular array geometries in 2D.
+
+- :py:func:`pyroomacoustics.beamforming.linear_2D_array`
+- :py:func:`pyroomacoustics.beamforming.circular_2D_array`
+- :py:func:`pyroomacoustics.beamforming.square_2D_array`
+- :py:func:`pyroomacoustics.beamforming.poisson_2D_array`
+- :py:func:`pyroomacoustics.beamforming.spiral_2D_array`
+
+
+Create the Room Impulse Response
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+At this point, the RIRs are simply created by invoking the ISM via
+:py:func:`pyroomacoustics.room.Room.image_source_model`. This function will
+generate all the images sources up to the order required and use them to
+generate the RIRs, which will be stored in the ``rir`` attribute of ``room``.
+The attribute ``rir`` is a list of lists so that the outer list is on microphones
+and the inner list over sources.
+
+.. code-block:: python
+
+    room.compute_rir()
+
+    # plot the RIR between mic 1 and source 0
+    import matplotlib.pyplot as plt
+    plt.plot(room.rir[1][0])
+    plt.show()
+
+
+Simulate sound propagation
+~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+By calling :py:func:`pyroomacoustics.room.Room.simulate`, a convolution of the
+signal of each source (if not ``None``) will be performed with the
+corresponding room impulse response. The output from the convolutions will be summed up
+at the microphones. The result is stored in the ``signals`` attribute of ``room.mic_array``
+with each row corresponding to one microphone.
+
+.. code-block:: python
+
+    room.simulate()
+
+    # plot signal at microphone 1
+    plt.plot(room.mic_array.signals[1,:])
+
+
+Example
+-------
+
+.. code-block:: python
+
+    import numpy as np
+    import matplotlib.pyplot as plt
+    import pyroomacoustics as pra
+
+    # Create a 4 by 6 metres shoe box room
+    room = pra.ShoeBox([4,6])
+
+    # Add a source somewhere in the room
+    room.add_source([2.5, 4.5])
+
+    # Create a linear array beamformer with 4 microphones
+    # with angle 0 degrees and inter mic distance 10 cm
+    R = pra.linear_2D_array([2, 1.5], 4, 0, 0.04) 
+    room.add_microphone_array(pra.Beamformer(R, room.fs))
+
+    # Now compute the delay and sum weights for the beamformer
+    room.mic_array.rake_delay_and_sum_weights(room.sources[0][:1])
+
+    # plot the room and resulting beamformer
+    room.plot(freq=[1000, 2000, 4000, 8000], img_order=0)
+    plt.show()
+
+'''
+
 
 from __future__ import print_function
 
@@ -20,9 +188,23 @@ from .c_package import libroom_available, CWALL, CROOM, libroom, c_wall_p, c_int
 
 class Room(object):
     '''
-    This class represents a room instance.
-    
-    A room instance is formed by wall instances. A MicrophoneArray and SoundSources can be added.
+    A Room object has as attributes a collection of
+    :py:obj:`pyroomacoustics.wall.Wall` objects, a
+    :py:obj:`pyroomacoustics.beamforming.MicrophoneArray` array, and a list of
+    :py:obj:`pyroomacoustics.soundsource.SoundSource`. The room can be two
+    dimensional (2D), in which case the walls are simply line segments. A factory method 
+    :py:func:`pyroomacoustics.room.Room.from_corners`
+    can be used to create the room from a polygon. In three dimensions (3D), the
+    walls are two dimensional polygons, namely a collection of points lying on a
+    common plane. Creating rooms in 3D is more tedious and for convenience a method
+    :py:func:`pyroomacoustics.room.Room.extrude` is provided to lift a 2D room
+    into 3D space by adding vertical walls and a parallel “ceiling” (see Figure
+    4b).
+
+    The Room is sub-classed by :py:obj:pyroomacoustics.room.ShoeBox` which
+    creates a rectangular (2D) or parallelepipedic (3D) room. Such rooms
+    benefit from an efficient algorithm for the image source method.
+
     
     :attribute walls: (Wall array) list of walls forming the room
     :attribute fs: (int) sampling frequency
@@ -124,7 +306,7 @@ class Room(object):
         absorption = np.array(absorption, dtype='float64')
         if (absorption.ndim == 0):
             absorption = absorption * np.ones(corners.shape[1])
-        elif (absorption.ndim > 1 and corners.shape[1] != len(absorption)):
+        elif (absorption.ndim >= 1 and corners.shape[1] != len(absorption)):
             raise ValueError('Arg absorption must be the same size as corners or must be a single value.')
         
         walls = []
@@ -268,10 +450,15 @@ class Room(object):
     def plot(self, img_order=None, freq=None, figsize=None, no_axis=False, mic_marker_size=10, **kwargs):
         ''' Plots the room with its walls, microphones, sources and images '''
     
-        import matplotlib
-        from matplotlib.patches import Circle, Wedge, Polygon
-        from matplotlib.collections import PatchCollection
-        import matplotlib.pyplot as plt
+        try:
+            import matplotlib
+            from matplotlib.patches import Circle, Wedge, Polygon
+            from matplotlib.collections import PatchCollection
+            import matplotlib.pyplot as plt
+        except ImportError:
+            import warnings
+            warnings.warn('Matplotlib is required for plotting')
+            return
 
         if (self.dim == 2):
             fig = plt.figure(figsize=figsize)
@@ -429,7 +616,13 @@ class Room(object):
         if self.rir is None:
             self.compute_rir()
 
-        import matplotlib.pyplot as plt
+        try:
+            import matplotlib.pyplot as plt
+        except ImportError:
+            import warnings
+            warnings.warn('Matplotlib is required for plotting')
+            return
+
         from . import utilities as u
 
         M = self.mic_array.M
@@ -448,6 +641,8 @@ class Room(object):
                         plt.xlabel('Time [s]')
                     else:
                         plt.xlabel('Normalized frequency')
+
+        plt.tight_layout()
 
 
     def add_microphone_array(self, micArray):